rename proto/private to proto/private_key
