+++ src/proto/private_key.ts
